v8_engine: Put executor_stress test in separate file
