+++ src/v/v8_engine/tests/executor_test.cc
@@ -195,55 +195,3 @@ SEASTAR_THREAD_TEST_CASE(stop_executor_with_finish_item_test) {
         }
     }
 }
-
-SEASTAR_THREAD_TEST_CASE(stress_test) {
-    struct task_for_test {
-        explicit task_for_test(char& is_finish)
-          : _is_finish(is_finish) {}
-
-        void operator()() { _is_finish = 1; }
-
-        void cancel() {}
-
-        void on_timeout() {}
-
-        char& _is_finish;
-    };
-
-    std::vector<size_t> queue_size = {1, 10, 1000};
-
-    for (auto size : queue_size) {
-        v8_engine::executor test_executor(ss::engine().alien(), 1, size);
-
-        std::unordered_map<size_t, char> finish_flags;
-        std::vector<ss::future<>> futures;
-
-        auto finish_time = ss::lowres_clock::now() + std::chrono::seconds(10);
-
-        size_t it_count = 0;
-
-        while (ss::lowres_clock::now() < finish_time) {
-            if (it_count % 5 == 0) {
-                char is_finish = 0;
-                auto fut = test_executor.submit(
-                  task_for_test(is_finish), std::chrono::milliseconds(100));
-                fut.get();
-                BOOST_REQUIRE_EQUAL(is_finish, 1);
-            } else {
-                finish_flags[it_count] = 0;
-                futures.emplace_back(test_executor.submit(
-                  task_for_test(finish_flags[it_count]),
-                  std::chrono::milliseconds(5000)));
-            }
-
-            ++it_count;
-        }
-
-        auto futures_res = ss::when_all(futures.begin(), futures.end()).get();
-        for (auto res : finish_flags) {
-            BOOST_REQUIRE_EQUAL(res.second, 1);
-        }
-
-        test_executor.stop().get();
-    }
-}
